change struct recvfrom_args and struct sendto_args to take a size_t as a length parameter

+++ sys/kern/uipc_syscalls.c
@@ -4,7 +4,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)uipc_syscalls.c	8.3 (Berkeley) 01/04/94
+ *	@(#)uipc_syscalls.c	8.4 (Berkeley) 02/21/94
  */
 
 #include <sys/param.h>
@@ -323,7 +323,7 @@ free1:
 struct sendto_args {
 	int	s;
 	caddr_t	buf;
-	int	len;
+	size_t	len;
 	int	flags;
 	caddr_t	to;
 	int	tolen;
@@ -556,7 +556,7 @@ bad:
 struct recvfrom_args {
 	int	s;
 	caddr_t	buf;
-	int	len;
+	size_t	len;
 	int	flags;
 	caddr_t	from;
 	int	*fromlenaddr;
